重构：数据库相关
1. 支持sqlite
2. 更换mysql驱动（Pomelo.EntityFrameworkCore.MySql）

BreakChange: 由于更换驱动，原有的迁移文件大部分报错，虽然可以手动修改。再加上新的sqlite也需要迁移，为保持同步，需要删库重新迁移

+++ src/Application.Core.EF/Entities/MakerCreatedataEntity.cs
@@ -2,13 +2,13 @@ namespace Application.EF.Entities;
 
 public partial class MakerCreatedataEntity
 {
-    public sbyte Id { get; set; }
+    public int Id { get; set; }
 
     public int Itemid { get; set; }
 
-    public sbyte ReqLevel { get; set; }
+    public short ReqLevel { get; set; }
 
-    public sbyte ReqMakerLevel { get; set; }
+    public short ReqMakerLevel { get; set; }
 
     public int ReqMeso { get; set; }
